added Database Context and launching procedure

+++ Models/UserModel.cs
@@ -1,7 +1,7 @@
 using Microsoft.AspNetCore.Identity;
 
 namespace TMonitBackend.Models{
-    public class User : IdentityUser<long>{
+    public class User : IdentityUser<int>{
         
     }
 }
